+++ src/felupe/math/_tensor.py
@@ -1481,3 +1481,55 @@ def equivalent_von_mises(A):
 
     devA = dev(A)
     return np.sqrt(3 / 2 * ddot(devA, devA))
+
+
+def inplane(A, vectors, **kwargs):
+    r"""Return the in-plane components of symmetric 2x2 or 3x3 tensors, where the planes
+    are defined by their standard unit vectors.
+
+    Parameters
+    ----------
+    A : ndarray of shape (N, N, ...)
+        Symmetric second-order tensors.
+    vectors : list of ndarray of shape (N, ...)
+        List of standard unit vectors of the planes.
+    **kwargs : dict, optional
+        Optional keyword-arguments for :func:`numpy.einsum`.
+
+    Notes
+    -----
+    The first two axes of the tensor and the first axis of each vector are the
+    tensor/vector dimensions and all remaining trailing axes are treated as batch
+    dimensions.
+
+    ..  math::
+
+        \boldsymbol{A}_{N-1} = \sum_{\alpha, \beta}
+            \left( \boldsymbol{E}_\alpha^T \boldsymbol{A} \boldsymbol{E}_\beta\right)
+            \boldsymbol{E}_\alpha \otimes \boldsymbol{E}_\beta
+
+    Returns
+    -------
+    ndarray of shape (N - 1, N - 1, ...)
+        The in-plane components of symmetric (N, N) tensors.
+
+    Examples
+    --------
+    >>> import felupe as fem
+    >>>
+    >>> A = np.array([[0, 3, 5], [3, 1, 4], [5, 4, 2]])
+    >>> vectors = [[1, 0, 0], [0, 1, 0]]
+    >>>
+    >>> A_inplane = fem.math.inplane(A, vectors)
+    >>> A_inplane
+    array([[0, 3],
+           [3, 1]])
+    """
+
+    vectors = np.array(vectors)
+
+    i, j = np.triu_indices(len(vectors))
+    ij = np.zeros((len(vectors), len(vectors)), dtype=int)
+    ij[i, j] = ij[j, i] = np.arange(len(i), dtype=int)
+
+    return np.einsum("ij...,ai...,aj...->a...", A, vectors[i], vectors[j], **kwargs)[ij]